Use collocated templates for LiveView and LiveComponents
Starting on live_view 0.10.0 we can use a simpler structure where live
views and components will automatically render a template with the same
name as the view/component if they're located in the same directory.
This makes unnecessary implementing render/1 in every component that
only calls render/3 passing the assigns.
For presentation functions, which normally are defined in the view, I
decided to move them to the component modules so there's no need to have
views. The only caveat is that these aren't imported in the templates
and we need to call them using the module (aliased).
