+++ lib/elixir_console_web/live/console_live.html.leex
@@ -1,3 +1,5 @@
+<% alias ElixirConsoleWeb.ConsoleLive.{CommandInputComponent, HistoryComponent, SidebarComponent} %>
+
 <div class="flex h-full flex-col sm:flex-row">
   <div class="flex-1 sm:h-full overflow-scroll">
     <div class="h-full flex flex-col">
