+++ lib/elixir_console_web/live/console_live/history_component.ex
@@ -5,10 +5,8 @@ defmodule ElixirConsoleWeb.ConsoleLive.HistoryComponent do
   """
 
   use Phoenix.LiveComponent
-
-  def render(assigns) do
-    Phoenix.View.render(ElixirConsoleWeb.Console.HistoryView, "index.html", assigns)
-  end
+  import Phoenix.HTML, only: [sigil_e: 2]
+  alias ElixirConsole.ContextualHelp
 
   def handle_event(
         "function_link_clicked",
@@ -22,4 +20,41 @@ defmodule ElixirConsoleWeb.ConsoleLive.HistoryComponent do
 
     {:noreply, socket}
   end
+
+  def format_command(command) do
+    for part <- split_command(command) do
+      case part do
+        {part, help_metadata} ->
+          render_command_inline_help(part, help_metadata)
+
+        part ->
+          part
+      end
+    end
+  end
+
+  defp split_command(command) do
+    ContextualHelp.compute(command)
+  end
+
+  defp render_command_inline_help(part, %{
+         type: function_or_operator,
+         func_name: func_name,
+         header: header,
+         docs: docs,
+         link: link
+       }) do
+    ~e{<span
+    phx-click="function_link_clicked"
+    phx-value-func_name="<%= func_name %>"
+    phx-value-header="<%= header %>"
+    phx-value-doc="<%= docs %>"
+    phx-value-link="<%= link %>"
+    phx-target="#commandOutput"
+    class="<%= inline_help_class_for(function_or_operator) %>"
+    ><%= part %></span>}
+  end
+
+  defp inline_help_class_for(:function), do: "text-green-400 cursor-pointer underline"
+  defp inline_help_class_for(:operator), do: "cursor-pointer"
 end
