+++ lib/elixir_console_web/live/console_live/sidebar_component.ex
@@ -6,8 +6,4 @@ defmodule ElixirConsoleWeb.ConsoleLive.SidebarComponent do
   """
 
   use Phoenix.LiveComponent
-
-  def render(assigns) do
-    Phoenix.View.render(ElixirConsoleWeb.ConsoleView, "sidebar.html", assigns)
-  end
 end
