+++ lib/elixir_console_web/live/console_live/sidebar_component.html.leex
