--- lib/elixir_console_web/views/console_view.ex
@@ -1,8 +1,0 @@
-defmodule ElixirConsoleWeb.ConsoleView do
-  use ElixirConsoleWeb, :view
-  import Phoenix.LiveView.Helpers, only: [live_component: 3]
-
-  alias ElixirConsoleWeb.ConsoleLive.{CommandInputComponent, HistoryComponent, SidebarComponent}
-
-  def print_prompt, do: "> "
-end
